Just return an int for the single-byte bearing

+++ CMPS10.h
@@ -16,7 +16,7 @@ class CMPS10 {
     public:
         CMPS10();
         CMPS10(int i2c_address);
-        int8_t bearing();
+        int bearing();
         int8_t pitch();
         int8_t roll();
     private:
